Доработан генетический алгоритм для задачи коммивояжера

+++ src/algorithms/constants.rs
@@ -1,3 +1,3 @@
 pub const ITERS_COUNT: u64 = 50;
 pub const ACTORS_COUNT: usize = 100;
-pub const SOLUTIONS_COUNT: u64 = 3;
+pub const SOLUTIONS_COUNT: usize = 3;
